+++ src/main/java/ladder/domain/ladder/DirectionStrategy.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 @FunctionalInterface
 public interface DirectionStrategy {
