+++ src/main/java/ladder/domain/ladder/RandomDirectionStrategy.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import java.util.Random;
